Add script for building live images in one swoop
To generate live images, several commands have to be executed in a
row. Having a script that abstracts these steps would simplify the
process.

This commit adds a script that performs all the steps that are
necessary to generate live images. The script allows the user to
specify the desired AlmaLinux version, desktop environment, and
image variant (full or mini), as well as the package cache directory
on the command line.